+++ app/src/main/java/com/jkdajac/cookingnote/salad/SaladActivity.kt
@@ -4,6 +4,8 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.animation.AnimationUtils
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.jkdajac.cookingnote.MainActivity
 import com.jkdajac.cookingnote.R
@@ -58,10 +60,23 @@ class SaladActivity : AppCompatActivity(), SaladAdapter.ViewHolder.ItemCallback 
         saladList.addAll(wordFromDb)
     }
     override fun deleteItem(index: Int) {
-        val salad = saladList[index]
-        saladDatabase.saladDao().deleteSalad(salad)
-        getData()
-        adapter.notifyDataSetChanged()
+
+        val addDialog = AlertDialog.Builder(this)
+        addDialog
+            .setMessage("Вы действительно хотите удалить запись?")
+            .setPositiveButton("Ok") { dialog, _ ->
+                val salad = saladList[index]
+                saladDatabase.saladDao().deleteSalad(salad)
+                getData()
+                adapter.notifyDataSetChanged()
+                Toast.makeText(this, "Запись удалена!", Toast.LENGTH_SHORT).show()
+                dialog.dismiss()
+            }
+            .setNegativeButton("Отмена") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
+            .show()
     }
